Migrate Neo4j sampler onto the SamplerExtensionPoint mechanism
