+++ roadrunner-sampler-neo4j/src/main/java/io/roadrunner/samplers/neo4j/Neo4jSampler.java
@@ -1,0 +1,55 @@
+/**
+ * Copyright 2024 Symentis.pl
+ * <p>
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * <p>
+ * http://www.apache.org/licenses/LICENSE-2.0
+ * <p>
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.roadrunner.samplers.neo4j;
+
+import io.roadrunner.api.events.SamplerResponse;
+import io.roadrunner.api.samplers.Sampler;
+import java.util.Map;
+import org.neo4j.driver.Driver;
+
+/**
+ * Extension-point methods class for the Neo4j sampler: {@link #query(String)} is bound from a
+ * CLI {@code query("RETURN 1")} expression via {@link io.roadrunner.samplers.spi.SamplerExtensionPoint}.
+ */
+public class Neo4jSampler {
+
+    private final Driver driver;
+
+    public Neo4jSampler(Driver driver) {
+        this.driver = driver;
+    }
+
+    public Sampler query(String cypher) {
+        return parameters -> {
+            var startTime = System.nanoTime();
+            try (var session = driver.session()) {
+                // Neo4j's Session.run accepts Map<String, Object>; SamplerParameters.asMap returns
+                // Map<String, ?>. Erasure makes the cast safe — Session.run is a read-only consumer
+                // of the map. See #137 for typed CSV values (Integer/Long/etc. instead of String).
+                @SuppressWarnings("unchecked")
+                var params = (Map<String, Object>) parameters.asMap();
+                var result = session.run(cypher, params);
+                return SamplerResponse.response(startTime, System.nanoTime(), result.consume());
+            } catch (Exception e) {
+                return SamplerResponse.error(startTime, System.nanoTime(), e.getMessage());
+            }
+        };
+    }
+
+    public void close() {
+        driver.close();
+    }
+}
